changed styles

+++ TimeTrackerApp/GraphQL/GraphQLQueries/AppMutation.cs
@@ -13,31 +13,38 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
 {
     public class AppMutation : ObjectGraphType
     {
-        public AppMutation(IHttpContextAccessor contextAccessor,ICalendarRepository calendarRepository, IAuthenticationTokenRepository authenticationTokenRepository, IRecordRepository recordRepository, IUserRepository userRepository, IVacationRepository vacationRepository, ISickLeaveRepository sickLeaveRepository)
+        public AppMutation(IHttpContextAccessor contextAccessor, ICalendarRepository calendarRepository,
+            IAuthenticationTokenRepository authenticationTokenRepository, IRecordRepository recordRepository,
+            IUserRepository userRepository, IVacationRepository vacationRepository,
+            ISickLeaveRepository sickLeaveRepository)
         {
             var authenticationService = new AuthenticationService(userRepository, authenticationTokenRepository);
-            
+
             Field<UserType, User>()
                 .Name("ChangedPrivelege")
                 .Argument<IntGraphType, int>("userId", "user id whose privelege will be update")
                 .Argument<IntGraphType, int>("privilegeValue", "value privilege")
                 .ResolveAsync(async context =>
                 {
-                    var userId = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserId").Value);
-                    if (userId==context.GetArgument<int>("userId"))
+                    var userId = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserId").Value);
+                    if (userId == context.GetArgument<int>("userId"))
                     {
                         throw new Exception("You cant change yourself privilege");
                     }
-                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserPrivilegesValue").Value);
-                    var result =userPermission & Convert.ToInt32(Privileges.EditUsers);
-                    if (!(result>0))
+
+                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserPrivilegesValue").Value);
+                    var result = userPermission & Convert.ToInt32(Privileges.EditUsers);
+                    if (!(result > 0))
                     {
                         throw new Exception("You dont have permission to edit users");
                     }
+
                     var valuePrivilege = context.GetArgument<int>("privilegeValue");
                     var user = await userRepository.ChangePrivelegeValueAsync(new User
                     {
-                        Id =userId,
+                        Id = userId,
                         PrivilegesValue = valuePrivilege
                     });
                     return user;
@@ -94,18 +101,20 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
             //             PrivilegesValue = valuePrivilegeUser +  addValuePrivilege
             //         });
             //     });
-            
+
             Field<UserType, User>()
                 .Name("CreateUser")
                 .Argument<NonNullGraphType<UserInputType>, User>("User", "User")
                 .ResolveAsync(async context =>
                 {
-                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserPrivilegesValue").Value);
+                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserPrivilegesValue").Value);
                     var result = userPermission & Convert.ToInt32(Privileges.CreateUsers);
-                    if (!(result>0))
+                    if (!(result > 0))
                     {
                         throw new Exception("You dont have permission to create users");
                     }
+
                     var user = context.GetArgument<User>("User");
                     return await userRepository.CreateAsync(user);
                 });
@@ -115,13 +124,20 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                 .Argument<NonNullGraphType<IdGraphType>, int>("Id", "User id")
                 .ResolveAsync(async context =>
                 {
-                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserPrivilegesValue").Value);
-                    var result =userPermission & Convert.ToInt32(Privileges.DeleteUsers);
-                    if (!(result>0))
+                    var userId = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserId").Value);
+                    if (userId == context.GetArgument<int>("userId"))
+                    {
+                        throw new Exception("You cant delete yourself");
+                    }
+                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserPrivilegesValue").Value);
+                    var result = userPermission & Convert.ToInt32(Privileges.DeleteUsers);
+                    if (!(result > 0))
                     {
                         throw new Exception("You dont have permission to delete users");
                     }
-                    
+
                     int id = context.GetArgument<int>("Id");
                     return await userRepository.RemoveAsync(id);
                 });
@@ -131,13 +147,13 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                 .Argument<NonNullGraphType<UserInputType>, User>("User", "User")
                 .ResolveAsync(async context =>
                 {
-                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserPrivilegesValue").Value);
-                    var result =userPermission & Convert.ToInt32(Privileges.EditUsers);
-                    if (!(result>0))
+                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserPrivilegesValue").Value);
+                    var result = userPermission & Convert.ToInt32(Privileges.EditUsers);
+                    if (!(result > 0))
                     {
                         throw new Exception("You dont have permission to edit users");
                     }
-                    
                     var user = context.GetArgument<User>("User");
                     return await userRepository.EditAsync(user);
                 });
@@ -167,21 +183,23 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                 .Argument<NonNullGraphType<RecordInputType>, Record>("Record", "Record")
                 .ResolveAsync(async context =>
                 {
-                    var record = context.GetArgument<Record>("Record"); 
-                    var userId = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserId").Value);
+                    var record = context.GetArgument<Record>("Record");
+                    var userId = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserId").Value);
                     if (record.EmployeeId == userId)
                     {
-                          return await recordRepository.EditAsync(record);
+                        return await recordRepository.EditAsync(record);
                     }
-                    
-                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims.First(x=>x.Type=="UserPrivilegesValue").Value);
-                    var result =userPermission & Convert.ToInt32(Privileges.EditUsers);
-                    if (!(result>0))
+
+                    var userPermission = int.Parse(contextAccessor.HttpContext.User.Identities.First().Claims
+                        .First(x => x.Type == "UserPrivilegesValue").Value);
+                    var result = userPermission & Convert.ToInt32(Privileges.EditUsers);
+                    if (!(result > 0))
                     {
                         throw new Exception("You dont have permission to edit records for another users");
                     }
+
                     return await recordRepository.EditAsync(record);
-                  
                 });
 
             Field<RecordType, Record>()
@@ -192,7 +210,7 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     int id = context.GetArgument<int>("Id");
                     return await recordRepository.RemoveAsync(id);
                 });
-            
+
             Field<VacationType, Vacation>()
                 .Name("CreateVacationRequest")
                 .Argument<NonNullGraphType<VacationInputType>, Vacation>("VacationRequest", "Vacation request")
@@ -201,7 +219,7 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     var vacationRequest = context.GetArgument<Vacation>("VacationRequest");
                     return await vacationRepository.CreateAsync(vacationRequest);
                 });
-            
+
             Field<VacationType, Vacation>()
                 .Name("EditVacationRequest")
                 .Argument<NonNullGraphType<VacationInputType>, Vacation>("VacationRequest", "Vacation request")
@@ -210,21 +228,22 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     var vacationRequest = context.GetArgument<Vacation>("VacationRequest");
                     return await vacationRepository.EditAsync(vacationRequest);
                 });
-            
+
             Field<VacationType, Vacation>()
                 .Name("DeleteVacationRequest")
                 .Argument<IntGraphType, int>("Id", "Vacation request")
                 .ResolveAsync(async context =>
                 {
                     int id = context.GetArgument<int>("Id");
-                    var model =  await vacationRepository.RemoveAsync(id);
+                    var model = await vacationRepository.RemoveAsync(id);
 
                     return model;
                 });
 
             Field<AuthTokenType, AuthenticationToken>()
                 .Name("CreateAuthenticationToken")
-                .Argument<NonNullGraphType<AuthTokenInputType>, AuthenticationToken>("AuthToken", "Authentication token")
+                .Argument<NonNullGraphType<AuthTokenInputType>, AuthenticationToken>("AuthToken",
+                    "Authentication token")
                 .ResolveAsync(async context =>
                 {
                     var authenticationToken = context.GetArgument<AuthenticationToken>("AuthToken");
@@ -233,7 +252,8 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
 
             Field<AuthTokenType, AuthenticationToken>()
                 .Name("EditAuthenticationToken")
-                .Argument<NonNullGraphType<AuthTokenInputType>, AuthenticationToken>("AuthToken", "Authentication token")
+                .Argument<NonNullGraphType<AuthTokenInputType>, AuthenticationToken>("AuthToken",
+                    "Authentication token")
                 .ResolveAsync(async context =>
                 {
                     var authenticationToken = context.GetArgument<AuthenticationToken>("AuthToken");
@@ -241,13 +261,13 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                 });
 
             Field<AuthTokenType, AuthenticationToken>()
-               .Name("DeleteAuthenticationToken")
-               .Argument<NonNullGraphType<IdGraphType>, int>("Id", "Authentication token id")
-               .ResolveAsync(async context =>
-               {
-                   int id = context.GetArgument<int>("Id");
-                   return await authenticationTokenRepository.RemoveAsync(id);
-               });
+                .Name("DeleteAuthenticationToken")
+                .Argument<NonNullGraphType<IdGraphType>, int>("Id", "Authentication token id")
+                .ResolveAsync(async context =>
+                {
+                    int id = context.GetArgument<int>("Id");
+                    return await authenticationTokenRepository.RemoveAsync(id);
+                });
 
             Field<AuthTokenType, AuthenticationToken>()
                 .Name("DeleteAuthenticationTokenByUserId")
@@ -266,9 +286,9 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                 {
                     string email = context.GetArgument<string>("Email");
                     string password = context.GetArgument<string>("Password");
-                    
+
                     try
-					{
+                    {
                         var authenticationServiceResponse = await authenticationService.Login(email, password);
                         var authenticationServiceApiResponse = new AuthResponse()
                         {
@@ -278,10 +298,10 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                         return authenticationServiceApiResponse;
                     }
                     catch (Exception exception)
-					{
+                    {
                         context.Errors.Add(new ExecutionError(exception.Message));
                         return new AuthResponse();
-					}
+                    }
                 });
 
             Field<AuthResponseType, AuthResponse>()
@@ -311,19 +331,20 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     var refreshToken = context.GetArgument<string>("RefreshToken");
                     try
                     {
-                        var authenticationServiceResponse = await authenticationService.Refresh(userId, accessToken, refreshToken);
+                        var authenticationServiceResponse =
+                            await authenticationService.Refresh(userId, accessToken, refreshToken);
                         var authenticationServiceApiResponse = new AuthResponse()
                         {
                             AccessToken = authenticationServiceResponse.AccessToken,
                             RefreshToken = authenticationServiceResponse.RefreshToken
                         };
                         return authenticationServiceApiResponse;
-                    } 
+                    }
                     catch (Exception exception)
-					{
+                    {
                         context.Errors.Add(new ExecutionError(exception.Message));
                         return new AuthResponse();
-					}
+                    }
                 });
 
             Field<CalendarDayType, CalendarDay>()
@@ -343,7 +364,7 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     var id = context.GetArgument<int>("Id");
                     return await calendarRepository.RemoveDayAsync(id);
                 });
-            
+
             Field<CalendarDayType, CalendarDay>()
                 .Name("EditCalendarDay")
                 .Argument<CalendarDayUpdateType, CalendarDay>("Day", "CalendarDay")
@@ -352,8 +373,8 @@ namespace TimeTrackerApp.GraphQL.GraphQLQueries
                     var day = context.GetArgument<CalendarDay>("Day");
                     return await calendarRepository.EditDayAsync(day);
                 });
-            
-            Field<VacationType,Vacation>()
+
+            Field<VacationType, Vacation>()
                 .Name("CreateVacation")
                 .Argument<VacationInputType, Vacation>("Vacation", "vacation arguments")
                 .ResolveAsync(async _ =>
